% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_connection}
\alias{check_connection}
\title{[INTERNAL] Check connection}
\usage{
check_connection(connection)
}
\arguments{
\item{connection}{[list] Connection to check. Created by \code{\link[DrDimont]{make_connection}}}
}
\value{
Character string vector containing error messages.
}
\description{
[INTERNAL] Checks if the data given to create an inter-layer connection is valid and has the
right input format
}
\keyword{internal}
