% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_drug_targets_in_layers}
\alias{check_drug_targets_in_layers}
\title{[INTERNAL] Check drug target and layer data}
\usage{
check_drug_targets_in_layers(drug_target_interactions, layers)
}
\arguments{
\item{drug_target_interactions}{[list] A named list of the drug interaction data. Created by
\code{\link[DrDimont]{make_drug_target}}}

\item{layers}{[list] List of layers to check. Individual layers are created by
\code{\link[DrDimont]{make_layer}} and need to be wrapped in a list.}
}
\value{
Character string vector containing error messages.
}
\description{
[INTERNAL] Checks if the parameters supplied in 'drug_target_interactions' makes
sense in the context of the defined layers.
}
\keyword{internal}
