% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_input}
\alias{check_input}
\title{Check pipeline input data for required format}
\usage{
check_input(layers, inter_layer_connections, drug_target_interactions)
}
\arguments{
\item{layers}{[list] List of layers to check. Individual layers were created by
\code{\link[DrDimont]{make_layer}} and need to be wrapped in a list.}

\item{inter_layer_connections}{[list] A list containing connections between layers. Each
connection was created by \code{\link[DrDimont]{make_connection}} and wrapped in a list.}

\item{drug_target_interactions}{[list] A named list of the drug interaction data. Created by
\code{\link[DrDimont]{make_drug_target}}}
}
\value{
Character string vector containing error messages.
}
\description{
Checks if input data is valid and formatted correctly. This function is a
wrapper for other check functions to be executed as first step of the DrDimont pipeline.
}
