% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_layer}
\alias{check_layer}
\title{[INTERNAL] Check layer input}
\usage{
check_layer(layer)
}
\arguments{
\item{layer}{[list] Named list of layer to check. Created by \code{\link[DrDimont]{make_layer}}}
}
\value{
Character string vector containing error messages.
}
\description{
[INTERNAL] Checks if the data used to create a network layer is valid and has the right
format
}
\keyword{internal}
