% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_sensible_connections}
\alias{check_sensible_connections}
\title{[INTERNAL] Check connection and layer data}
\usage{
check_sensible_connections(connection, layers)
}
\arguments{
\item{connection}{[list] Connection to check. Created by \code{\link[DrDimont]{make_connection}}}

\item{layers}{[list] List of layers to check. Individual layers are created by
\code{\link[DrDimont]{make_layer}} and need to be wrapped in a list.}
}
\value{
Character string vector containing error messages.
}
\description{
[INTERNAL] Checks if the connection defined in 'connection' makes sense in
context of the defined layers.
}
\keyword{internal}
