% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_reduction.R
\name{chunk}
\alias{chunk}
\title{[INTERNAL] Create chunks from a vector for parallel computing}
\source{
https://stackoverflow.com/questions/3318333/split-a-vector-into-chunks
}
\usage{
chunk(x, chunk_size)
}
\arguments{
\item{x}{Vector}

\item{chunk_size}{[int] Length of chunks}
}
\value{
A list of chunks of length chunk_size
}
\description{
[INTERNAL] Create chunks from a vector for parallel computing
}
