% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_reduction.R
\name{chunk_2gether}
\alias{chunk_2gether}
\title{[INTERNAL] Create chunks from two vectors for parallel computing}
\source{
modified from: https://stackoverflow.com/questions/3318333/split-a-vector-into-chunks
}
\usage{
chunk_2gether(x, y, chunk_size)
}
\arguments{
\item{x, y}{Vectors}

\item{chunk_size}{[int] Length of chunks}
}
\value{
A list of lists. Each second level list contains a list of chunks of length chunk_size of each
input vector.
}
\description{
[INTERNAL] Create chunks from two vectors for parallel computing
}
\keyword{internal}
