% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_reduction.R
\name{corPvalueStudentParallel}
\alias{corPvalueStudentParallel}
\title{[INTERNAL] Compute p-values for upper triangle of correlation matrix in parallel}
\usage{
corPvalueStudentParallel(adjacency_matrix, number_of_samples, chunk_size)
}
\arguments{
\item{adjacency_matrix}{[matrix] Adjacency matrix of correlations computed using \code{\link[WGCNA]{cor}} in
\code{\link[DrDimont]{compute_correlation_matrices}}}

\item{number_of_samples}{[matrix] Matrix of number of samples used in computation of each correlation value. Computed applying
\code{\link[DrDimont]{sample_size}}}

\item{chunk_size}{[int] Smallest unit of work in parallel computation (number of p-values to compute)}
}
\value{
Vector of p-values for upper triangle
}
\description{
[INTERNAL] Compute p-values for upper triangle of correlation matrix in parallel
}
\keyword{internal}
