% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentations.R
\docType{data}
\name{drug_gene_interactions}
\alias{drug_gene_interactions}
\title{Drug-gene interactions}
\format{
A dataframe with 4 columns.
\describe{
  \item{gene_name}{Gene names of targeted protein-coding genes.}
  \item{drug_name}{Drug-names with known interactions.}
  \item{drug_chembl_id}{ChEMBL ID of drugs.}
}
}
\source{
The Drug Gene Interaction Database: \url{https://dgidb.org/}

ChEMBL IDs: \url{https://www.ebi.ac.uk/chembl}
}
\usage{
drug_gene_interactions
}
\description{
Dataframe providing interactions of drugs with genes. The data was downloaded from
The Drug Gene Interaction Database.
}
\keyword{datasets}
