% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{get_layer_setting}
\alias{get_layer_setting}
\title{[INTERNAL] Get layer (and group) settings}
\usage{
get_layer_setting(layer, group, settings, setting_name)
}
\arguments{
\item{layer}{[list] A network layer created by \code{\link[DrDimont]{make_layer}}}

\item{group}{[string] A network group}

\item{settings}{[list] Named list of settings created by \code{\link[DrDimont]{drdimont_settings}}}

\item{setting_name}{[string] String indicating the setting to return.}
}
\value{
Setting value(s) for this layer (and group)
}
\description{
Returns specified setting for a specific network layer (and group).
}
\keyword{internal}
