% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_combination.R
\name{inter_layer_edgelist_by_id}
\alias{inter_layer_edgelist_by_id}
\title{[INTERNAL] Inter layer connections by identifiers}
\usage{
inter_layer_edgelist_by_id(annotation_A, annotation_B, connection, weight = 1)
}
\arguments{
\item{annotation_A, annotation_B}{[data.frame] Annotation tables specifying the identifiers of
the nodes of a network}

\item{connection}{[string] String of identifier to connect on}

\item{weight}{[int|vector] Integer or vector specifying the weight of the inter-layer connections.}
}
\value{
Data frame with columns from, to and weight
}
\description{
[INTERNAL] Returns an edge list defining the connections between two layers of the network.
}
