% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_score.R
\name{load_interaction_score_output}
\alias{load_interaction_score_output}
\title{[INTERNAL] Loads output of python script for interaction score calculation}
\usage{
load_interaction_score_output(saving_path, graphB_null)
}
\arguments{
\item{saving_path}{[string] Path to save intermediate output of DrDimont's functions. Default is current working directory.
Directory to use for writing intermediate data when passing input and output between Python and R.
Directory to load python output from}

\item{graphB_null}{[bool] Specifying if graphB of `groupB` is given (FALSE) or not (TRUE). (default: FALSE)}
}
\value{
A named list (elements `groupA` and `groupB`). Each element contains an iGraph object
containing the interaction score as edge attribute.
}
\description{
[INTERNAL] Loads data generated by \code{\link[DrDimont]{calculate_interaction_score}}. Python output files are
graphs in `gml` format for each of both groups.
}
