% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{make_drug_target}
\alias{make_drug_target}
\title{Reformat drug-target-interaction data}
\usage{
make_drug_target(target_molecules, interaction_table, match_on)
}
\arguments{
\item{target_molecules}{[string] Name of layer containing the drug targets. This name has to match the
corresponding named item in the list of layers supplied to \code{\link[DrDimont]{run_pipeline}}.}

\item{interaction_table}{[data.frame] Has to contain two columns. A column called `drug_name` containing
names or identifiers of drugs. And a column with a name that matches an identifier in the layer supplied
in `target_molecules`. Additional columns will be ignored in the pipeline.
For example, if drugs target proteins and an identifier called `ncbi_id` was supplied in layer creation of
the protein layer (see \code{\link[DrDimont]{make_layer}}), this column should be called
`ncbi_id` and contain the corresponding IDs of protein-drug targets. Any other ID present in
the constructed layer could also be used.}

\item{match_on}{[string] Column name of the dataframe supplied in `interaction_table` that is used for
matching drugs and target nodes in the graph (e.g. `ncbi_id`).}
}
\value{
Named list of the input parameters in input format of \code{\link[DrDimont]{run_pipeline}}.
}
\description{
Function to transform input data to required input format for
\code{\link[DrDimont]{run_pipeline}}. Here the data that is needed to define drug-target interactions is
formatted. When the reformatted output is passed to \code{\link[DrDimont]{run_pipeline}} as
\code{drug_target_interactions} argument, the differential integrated drug response score can be
calculated for all the supplied drugs in \code{interaction_table}.
}
\examples{
data(drug_gene_interactions)

example_drug_target_interactions <- make_drug_target(target_molecules='protein',
                                        interaction_table=drug_gene_interactions,
                                        match_on='gene_name')

}
