% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentations.R
\docType{data}
\name{metabolite_protein_interactions}
\alias{metabolite_protein_interactions}
\title{Metabolite protein interaction data}
\format{
A dataframe with 3 columns.
\describe{
  \item{pubchem_id}{Pubchem IDs defining interacting metabolites}
  \item{gene_name}{gene names defining interacting proteins}
  \item{combined_score}{Score describing the strength of metabolite-protein interaction}
}
}
\source{
STITCH DB: \url{https://stitch-db.org/}

Pubchem IDs: \url{https://pubchem.ncbi.nlm.nih.gov}

STRING DB: \url{https://string-db.org/}
}
\usage{
metabolite_protein_interactions
}
\description{
Dataframe providing interactions of metabolites and proteins. The data was taken
from the STITCH Database.
}
\keyword{datasets}
