% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_cluster}
\alias{set_cluster}
\title{[INTERNAL] Create and register cluster}
\usage{
set_cluster(n_threads)
}
\arguments{
\item{n_threads}{[int] Number of nodes in the cluster}
}
\value{
No return value, called internally to create cluster
}
\description{
[INTERNAL] Helper function to create and register a cluster for parallel
computation of p-value reduction
}
