% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{shutdown_cluster}
\alias{shutdown_cluster}
\title{[INTERNAL] Shutdown cluster and remove corresponding connections}
\usage{
shutdown_cluster()
}
\value{
No return value, called internally to shutdown cluster
}
\description{
[INTERNAL] Run this if the pipeline fails during parallel 
computation to clean the state. If a cluster is registered, this functions 
stops it and removes corresponding connections. Ignores errors. Has no effect 
if no cluster is registered.
}
