% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_targets.R
\name{target_edge_list}
\alias{target_edge_list}
\title{[INTERNAL] Get edges adjacent to target nodes}
\usage{
target_edge_list(graph, target_nodes, group)
}
\arguments{
\item{graph}{[igraph] Combined graph (iGraph graph object) for a specific group}

\item{target_nodes}{[data.frame] Has column `node_id` (unique node IDs in the iGraph graph
object that are targeted by drugs) and columns `groupA` and `groupB` (bool values specifying
whether the node is contained in the combined graph of the group)}

\item{group}{[string] Indicates which group `groupA` or `groupB` is analyzed}
}
\value{
An edge list as a data frame.
}
\description{
[INTERNAL] Based on the supplied graph and target nodes this function returns a
list of edges that are directly adjacent to target nodes. These edges can be used for further
computation to compute the integrated interaction scores and differential scores in the networks.
}
