% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_score.R
\name{write_interaction_score_input}
\alias{write_interaction_score_input}
\title{[INTERNAL] Write edge lists and combined graphs to files}
\usage{
write_interaction_score_input(
  combined_graphs,
  drug_target_edgelists,
  saving_path
)
}
\arguments{
\item{combined_graphs}{[list] A named list (elements `groupA` and `groupB`). Each element contains the
entire combined network (layers + inter-layer connections) as iGraph graph object.}

\item{drug_target_edgelists}{[list] A named list (elements `groupA` and `groupB`). Each element
contains the list of edges to be considered in the interaction score calculation as data frame (columns `from`, `to` and `weight`)}

\item{saving_path}{[string] Path to save intermediate output of DrDimont's functions.
Default is current working directory.}
}
\value{
No return value, used internally
}
\description{
[INTERNAL] Writes the combined graphs and the drug target edge lists to files for
passing them to the python interaction score script.
Graphs are saved as `gml` file. Edge lists are saved as `tsv` file.
}
