% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{return_errors}
\alias{return_errors}
\title{Return detected errors}
\usage{
return_errors(errors)
}
\arguments{
\item{errors}{[string] Character string vector containing error messages.}
}
\value{
No return value, writes error messages to console
}
\description{
Throws an error in case errors have been passed to the function. Messages
describing the detected errors are printed.
}
\examples{
layer <- DrDimont::layers_example[[2]]
return_errors(check_layer(layer))

}
